+++ src/models/StyledComponent.js
@@ -2,7 +2,7 @@
 
 import validAttr from '@emotion/is-prop-valid'
 import hoist from 'hoist-non-react-statics'
-import React, { Component, createElement } from 'react'
+import React, { createElement, PureComponent } from 'react'
 import ComponentStyle from './ComponentStyle'
 import createWarnTooManyClasses from '../utils/createWarnTooManyClasses'
 import determineTheme from '../utils/determineTheme'
@@ -58,7 +58,7 @@ const warnInnerRef = once(() =>
 )
 
 // $FlowFixMe
-class BaseStyledComponent extends Component<*> {
+class BaseStyledComponent extends PureComponent<*> {
   renderOuter: Function
   renderInner: Function
   styleSheet: ?StyleSheet